+++ src/main/java/com/telepathicgrunt/blame/mixin/SingleJigsawPieceAccessor.java
@@ -3,9 +3,12 @@ package com.telepathicgrunt.blame.mixin;
 import com.mojang.datafixers.util.Either;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.gen.feature.jigsaw.SingleJigsawPiece;
+import net.minecraft.world.gen.feature.template.StructureProcessorList;
 import net.minecraft.world.gen.feature.template.Template;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.function.Supplier;
 
 /* @author - TelepathicGrunt
  *
@@ -18,4 +21,7 @@ public interface SingleJigsawPieceAccessor {
 
     @Accessor("template")
     Either<ResourceLocation, Template> blame_getTemplateRL();
+
+    @Accessor("processors")
+    Supplier<StructureProcessorList> blame_getProcessors();
 }