+++ src/main/java/com/telepathicgrunt/blame/mixin/WorldEntitySpawnerMixin.java
@@ -2,6 +2,7 @@ package com.telepathicgrunt.blame.mixin;
 
 import net.minecraft.entity.EntityClassification;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.ChunkPos;
 import net.minecraft.world.IServerWorld;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.MobSpawnInfo;
@@ -20,6 +21,7 @@ import java.util.List;
 import java.util.Random;
 
 import static com.telepathicgrunt.blame.main.WorldEntitySpawnerBlame.addMobCrashDetails;
+import static com.telepathicgrunt.blame.main.WorldEntitySpawnerBlame.addMobGroupCrashDetails;
 
 /* @author - TelepathicGrunt
  *
@@ -47,6 +49,28 @@ public class WorldEntitySpawnerMixin {
                                                         int chunkX, int chunkZ, Random random,
                                                         CallbackInfo ci, MobSpawnInfo mobspawninfo,
                                                         List<MobSpawnInfo.Spawners> list) {
-        addMobCrashDetails(serverWorld, EntityClassification.CREATURE, chunkX, chunkZ, biome, list);
+        addMobCrashDetails(serverWorld, EntityClassification.CREATURE, new ChunkPos(chunkX, chunkZ).getWorldPosition(), biome, list);
+    }
+
+
+    @Inject(method = "getRandomSpawnMobAt(Lnet/minecraft/world/server/ServerWorld;Lnet/minecraft/world/gen/feature/structure/StructureManager;Lnet/minecraft/world/gen/ChunkGenerator;Lnet/minecraft/entity/EntityClassification;Ljava/util/Random;Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/MobSpawnInfo$Spawners;",
+            at = @At(value = "RETURN"),
+            locals = LocalCapture.CAPTURE_FAILHARD)
+    private static void blame_checkIfMobSpawnWillCrash3(ServerWorld serverWorld, StructureManager structureManager,
+                                                        ChunkGenerator chunkGenerator, EntityClassification entityClassification,
+                                                        Random random, BlockPos pos, CallbackInfoReturnable<MobSpawnInfo.Spawners> cir,
+                                                        Biome biome) {
+        addMobGroupCrashDetails(serverWorld, entityClassification, pos, biome, cir.getReturnValue());
+    }
+
+    @Inject(method = "spawnMobsForChunkGeneration(Lnet/minecraft/world/IServerWorld;Lnet/minecraft/world/biome/Biome;IILjava/util/Random;)V",
+            at = @At(value = "INVOKE", ordinal = 0, remap = false, target = "Ljava/util/Random;nextInt(I)I"),
+            locals = LocalCapture.CAPTURE_FAILHARD)
+    private static void blame_checkIfMobSpawnWillCrash4(IServerWorld serverWorld, Biome biome,
+                                                        int chunkX, int chunkZ, Random random,
+                                                        CallbackInfo ci, MobSpawnInfo mobspawninfo,
+                                                        List<MobSpawnInfo.Spawners> list, int i, int j,
+                                                        MobSpawnInfo.Spawners spawner) {
+        addMobGroupCrashDetails(serverWorld, EntityClassification.CREATURE, new ChunkPos(chunkX, chunkZ).getWorldPosition(), biome, spawner);
     }
 }